Download RedWolf XML from remote only if not available in local

+++ redwolf.cpp
@@ -7,12 +7,14 @@
 #include <fstream>
 #include <libxml/parser.h>
 #include <libxml/tree.h>
+#include <sys/stat.h>
 
 using namespace std;
 
 string GetEnv ( const string );
 string SplitString ( string, string, int );
 bool DirectoryExists( const char* );
+bool FileExists( const char* );
 string ToUpper( string );
 string GetIncremental( string );
 void UpdateXML( string, string, string, string );
@@ -22,7 +24,7 @@ int main()
 {
     string CMD;
     string OUT,TARGET_ARCH,TW_DEVICE_VERSION,TARGET_PRODUCT,BUILD_NUMBER;
-    string RW_BUILD,RW_WORK,RW_DEVICE,RW_OUT_NAME;
+    string RW_BUILD,RW_WORK,RW_DEVICE,RW_OUT_NAME,RW_XML_PATH;
 
     string RW_VENDOR = "vendor/redwolf";
 
@@ -89,6 +91,7 @@ int main()
     RW_WORK = OUT + "/RW_AIK";
     RW_DEVICE = SplitString(TARGET_PRODUCT,"_",1);
     RW_OUT_NAME = "RedWolf-" + RW_BUILD + "-" + RW_DEVICE;
+    RW_XML_PATH = "redwolf.xml";
 
 
     if(DirectoryExists(RW_WORK.c_str())) {
@@ -122,20 +125,26 @@ int main()
 
 
     if(TW_DEVICE_VERSION != "") {
-        cout << BLUE << "-- Downloading RedWolf XML from remote" << NC << endl;
-        CMD = "curl -o " + RW_WORK + "/redwolf.xml https://redwolfrecovery.github.io/redwolf.xml > /dev/null 2>&1";
-        system(CMD.c_str());
+        if(!FileExists(RW_XML_PATH.c_str())) {
+            cout << BLUE << "-- RedWolf XML not found. Downloading from remote" << NC << endl;
+            CMD = "curl -o " + RW_WORK + "/redwolf.xml https://redwolfrecovery.github.io/redwolf.xml > /dev/null 2>&1";
+            system(CMD.c_str());
+        }else{
+            cout << BLUE << "-- RedWolf XML found." << NC << endl;
+            CMD = "cp " + RW_XML_PATH + " " + RW_WORK + "/redwolf.xml";
+            system(CMD.c_str());
+        }
         cout << BLUE << "-- Reading build number from recovery" << NC << endl;
         BUILD_NUMBER = GetIncremental(RW_WORK + "/ramdisk/default.prop");
         cout << BLUE << "-- Updating build number in XML" << NC << endl;
-        UpdateXML(RW_WORK + "/redwolf.xml", OUT + "/redwolf.xml", RW_DEVICE, BUILD_NUMBER);
+        UpdateXML( RW_WORK + "/redwolf.xml", RW_XML_PATH, RW_DEVICE, BUILD_NUMBER);
     }
 
     cout << RED << "--------------------Finished making RedWolf---------------------" << NC << endl;
     cout << GREEN << "RedWolf image: ${OUT}/" << RW_OUT_NAME << ".img" << NC << endl;
     cout << GREEN << "          MD5: ${OUT}/" << RW_OUT_NAME << ".img.md5" << NC << endl;
     if(TW_DEVICE_VERSION != ""){
-    cout << GREEN << "          XML: ${OUT}/redwolf.xml" << NC << endl;}
+    cout << GREEN << "          XML: redwolf.xml" << NC << endl;}
     cout << RED << "================================================================" << NC << endl;
 
     return 0;
@@ -183,6 +192,11 @@ bool DirectoryExists( const char* pzPath ) {
     }
 
     return bExists;
+}
+
+bool FileExists( const char* filePath) {
+  struct stat buffer;
+  return (stat (filePath, &buffer) == 0); 
 }
 
 string ToUpper( string str ) {
